refeactor: car 생성 기능 Cars 객체 안으로 이동
