+++ src/test/java/domain/RacingGameTest.java
@@ -21,7 +21,7 @@ public class RacingGameTest {
                 new Car("kjm"),
                 new Car("bjs"),
                 new Car("honux")));
-        racingGame = new RacingGame(cars);
+        racingGame = new RacingGame();
     }
 
     @DisplayName("4이상의 숫자를 받은 차만 1칸 전진")
@@ -33,27 +33,12 @@ public class RacingGameTest {
         randomNumbers.add(4);
         randomNumbers.add(9);
 
-        racingGame.playGame(randomNumbers);
+        racingGame.playGame(cars, randomNumbers);
 
         assertThat(cars).isEqualTo(new Cars(Arrays.asList(new Car("kks", 0),
                 new Car("kjm", 0),
                 new Car("bjs", 1),
                 new Car("honux", 1))));
     }
-
-    @Test
-    public void getWinnerTest() {
-        List<Integer> randomNumbers = new ArrayList<>();
-        randomNumbers.add(0);
-        randomNumbers.add(3);
-        randomNumbers.add(4);
-        randomNumbers.add(9);
-
-        racingGame.playGame(randomNumbers);
-
-        assertThat(racingGame.getWinners()).hasSize(2);
-        assertThat(racingGame.getWinners()).contains("bjs", "honux");
-    }
-
 
 }
